Refactored option names in dictionary validator
To make things clearer, and, to make things consistent with the list validator that will be added next, universalConstraints has been re-named to valueConstraints, and keyConstraints to mapConstraints. This makes it possible for a new keyCosntraints option to be added which will apply valdidations to the key-part of the key-value pairs.